fix: 消除 MSVC/MinGW/macOS 编译告警和错误
- command_pipeline.hpp: if constexpr 加 else 块消除 C4702
- zlib_engine.cpp: size_t 到 uInt 加 static_cast 消除 C4267
- lock_free.hpp: #pragma 抑制 C4324 alignas 填充告警
- format.hpp: if constexpr 加 else 块消除 C4702
- map.hpp: destroy_at + construct_at 替代 = {}，修复 libc++ pair+unique_ptr 歧义
- json_simd.cpp: switch 改 if-else 链消除 C4065，backend 变量加守卫消除 C4189
- logger.cpp: if 改 if constexpr 消除 C4127
- pool.hpp: #pragma 抑制 C4324 alignas 填充告警
- crash_handler.hpp: #pragma comment 加 _MSC_VER 保护
- file_lock.hpp: LONGLONG 到 off_t 加 static_cast 消除 C4244
- platform.cpp: fgets size_t 到 int 加 static_cast 消除 C4267
- workflow_tools.cpp: 删除 catch 后不可达 return 消除 C4702
- terminal_io.hpp: #pragma 抑制 C4324 alignas 填充告警
- parser.cpp: ::tolower 改 lambda 消除 C4244
- handler.cpp: ::tolower 改 lambda 消除 C4244
- tests/CMakeLists.txt: -Wl,--stack 仅用于 MinGW，MSVC 用 /STACK:
- test_framework.hpp: while(0) 改 while((void)0,0) 消除 C4127
- test_json.cpp: int sum 改 int64_t 匹配 as_int() 返回类型
- test_net.cpp: 异常日志打印 + 消除 C4101

+++ src/base/platform/platform.cpp
@@ -251,7 +251,7 @@ process::ExecuteResult process::execute(const std::string& command) {
     }
 
     std::array<char, 4096> buffer;
-    while (fgets(buffer.data(), buffer.size(), pipe)) {
+    while (fgets(buffer.data(), static_cast<int>(buffer.size()), pipe)) {
         result.stdout_output += buffer.data();
     }
